ROCANA-9231: Fork PerFieldPostingsFormat (part 2 of 2)
Add javadocs and remove trailing whitespace

This must be a separate commit, not squashed. If we squash it the changes to TestRocanaPerFieldPostingsFormat.java will be too much and git will think we've deleted + added a new file rather than renamed and modified a file. That means we won't be able to do a diff from the current version of this file to the very first version committed (which is the pristine code, unmodified, from Lucene). That diff is valuable as it shows precisely what's changed in our fork.

Unfortunately git uses heuristics to determine when a file's been renamed and there's so many changes to this file it breaks that algorithm. For more details: https://git.wiki.kernel.org/index.php/GitFaq#Why_does_Git_not_.22track.22_renames.3F